requirements.txt
# lora module (RFM9X)
RPI.GPIO
adafruit-blinka
adafruit-circuitpython-rfm9x

#bme
adafruit-circuitpython-bme280

#parsing
crc8

# storing datas
paho-mqtt
influxdb

# ePaper
numpy
spidev
Pillow
matplotlib
